+++ web-client/www/src/js/app/stream/stream.viewer.js
@@ -1,4 +1,4 @@
-import { LargePacket } from "../utils/packet.js";
+import { LargePacket } from "../utils/packet.large.js";
 import { Demultiplexer } from "./stream.multiplexing.js";
 import Codec from "../utils/codec.js";
 import { CryptoPacket } from "../utils/packet.crypto.js";
